# docs/swagger.yaml
definitions:
  R.ResponseEntity:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  entity.UserEntity:
    properties:
      id:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  gin.H:
    additionalProperties: true
    type: object
...
info:
  contact:
    email: zengtao0618@163.com
    name: 曾涛
  description: 监控微服务系统中各个微服务实例的运行时状态以及所在容器的性能参数
  termsOfService: http://swagger.io/terms/
  title: FFR4MS Docker监控模块Fr-Docker
  version: "1.0"
paths:
...
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: 登录用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.UserEntity'
      produces:
      - application/json
      responses:
        "200":
          description: 返回登录Token
          schema:
            allOf:
            - $ref: '#/definitions/R.ResponseEntity'
            - properties:
                data:
                  $ref: '#/definitions/gin.H'
              type: object
        "400":
          description: 返回失败信息
          schema:
            $ref: '#/definitions/R.ResponseEntity'
      security:
      - ApiKeyAuth: []
      summary: 用户登录
      tags:
      - 用户操作
  /user/logout:
    post:
      description: 退出登录
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回登录Token
          schema:
            allOf:
            - $ref: '#/definitions/R.ResponseEntity'
            - properties:
                data:
                  $ref: '#/definitions/gin.H'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 退出登录
      tags:
      - 用户操作
...
swagger: "2.0"